Reclassification of customize data
Decided to get rid of customize in the repo. The hashes it accumulates
will cause issues mocing between machines, and it's a magnet for api
keys and passwords. I'm also not going to hardcode font settings in my
config for now, which is the main thing customize is still doing for me
right this moment.